Change modifications to topcat.properties and topcat.json
In topcat.properties, define LILS as the only facility, set its icatUrl
and idsUrl to icat_url, and unset any downloadType definitions (so that
they default to idsUrl).
In topcat.json, replace all instances of "http://localhost:8080" with
icat_url. This assumes that the same url can be used for topcatUrl,
icatUrl and the various idsUrls.